chore: refactor migrate module cmds
