+++ cmd/poktrolld/cmd/commands.go
@@ -26,7 +26,6 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/pokt-network/poktroll/app"
-	"github.com/pokt-network/poktroll/cmd/poktrolld/cmd/migrate"
 )
 
 func initRootCmd(
@@ -53,7 +52,6 @@ func initRootCmd(
 		queryCommand(),
 		txCommand(),
 		keys.Commands(),
-		migrate.MigrateCmd(),
 	)
 }
