+++ x/migration/module/cmd/errors.go
@@ -1,4 +1,4 @@
-package migrate
+package cmd
 
 import sdkerrors "cosmossdk.io/errors"
