+++ x/migration/module/cmd/migrate_test.go
@@ -1,4 +1,4 @@
-package migrate
+package cmd
 
 import (
 	"fmt"
