+++ x/migration/module/module.go
@@ -17,11 +17,13 @@ import (
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
+	"github.com/spf13/cobra"
 
 	// this line is used by starport scaffolding # 1
 
 	modulev1 "github.com/pokt-network/poktroll/api/poktroll/migration/module"
 	"github.com/pokt-network/poktroll/x/migration/keeper"
+	"github.com/pokt-network/poktroll/x/migration/module/cmd"
 	"github.com/pokt-network/poktroll/x/migration/types"
 )
 
@@ -85,6 +87,11 @@ func (AppModuleBasic) RegisterGRPCGatewayRoutes(clientCtx client.Context, mux *r
 	if err := types.RegisterQueryHandlerClient(context.Background(), mux, types.NewQueryClient(clientCtx)); err != nil {
 		panic(err)
 	}
+}
+
+// TODO_IN_THIS_COMMIT: godoc...
+func (AppModuleBasic) GetTxCmd() *cobra.Command {
+	return cmd.TxCommands()
 }
 
 // ----------------------------------------------------------------------------
